Fix problema con el monto bigdecimal

+++ src/main/java/org/jucamdonaciones/swgdsjucambackend/model/Donacion.java
@@ -1,7 +1,18 @@
 package org.jucamdonaciones.swgdsjucambackend.model;
 
-import jakarta.persistence.*;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EnumType;
+import jakarta.persistence.Enumerated;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "donaciones")
@@ -16,7 +27,7 @@ public class Donacion {
     private Donador donador;
 
     @Column(nullable = false, precision = 10, scale = 2)
-    private Double monto;
+    private BigDecimal monto;
 
     @Column(nullable = false)
     private LocalDateTime fechaDonacion = LocalDateTime.now();
@@ -67,11 +78,11 @@ public class Donacion {
         this.donador = donador;
     }
 
-    public Double getMonto() {
+    public BigDecimal getMonto() {
         return monto;
     }
 
-    public void setMonto(Double monto) {
+    public void setMonto(BigDecimal monto) {
         this.monto = monto;
     }
